reverted some changes

+++ include/label.h
@@ -5,10 +5,10 @@
 
 typedef struct label_s
 {
-    uint64_t left,right;
+    uint32_t left,right;
 }Label;
 //Create a new label
-Label label_init(uint64_t l,uint64_t r);
+Label label_init(uint32_t l,uint32_t r);
 //see if l1 is included in l2
 bool label_include(Label l1,Label l2);
